+++ lib/services/directory_service.dart
@@ -1,8 +1,13 @@
 import 'dart:io';
 
 import 'package:dcli/dcli.dart' as dcli;
+import 'package:emily/services/input_service.dart';
+
+import 'package:mason_logger/mason_logger.dart';
 
 import 'package:emily/utils/console.dart';
+import 'package:emily/constants/constants.dart';
+import 'package:emily/utils/validator.dart';
 
 /// This class provides functions to copy directories, create a folder structure and delete files.
 class DirectoryService {
@@ -38,6 +43,30 @@ class DirectoryService {
         Console.writeLine('Already exists: ${directory.path}');
       }
     }
+  }
+
+  static String choosePath() {
+    final logger = Logger();
+    final currentPath = Directory.current.path;
+
+    final result = logger.chooseOne(
+        Constants.kNeedSpecifyPath,
+        choices: <String?>[
+          Constants.kYes,
+          Constants.kNo
+        ],
+        defaultValue: Constants.kNo
+    );
+
+    if (result == Constants.kYes) {
+      return InputService.getValidatedInput(
+          consoleMessage: Constants.kEnterPath,
+          errorMessage: Constants.kInvalidPath,
+          functionValidator: Validator.isValidPath
+      ) ?? currentPath;
+    }
+
+    return currentPath;
   }