+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+## 0.1.0
+- create and Connect commands have been changed.
+- added possibility to generate keystore file and connect it to the project.
+
 ## 0.0.1
 - added possibility to create a Flutter project template.
 - added the ability to connect a remote Github repository.